+++ lib/scale_page_view.dart
@@ -20,7 +20,8 @@ class ScalePageView extends StatefulWidget {
       this.pageRatio: 0.5,
       this.scaleRatio: 0.2,
       this.opacityRatio: 0.5,
-      this.paddingTB: 23.0})
+      this.paddingTB: 23.0,
+      this.physics})
       : assert(children != null && children.length > 0),
         assert(pageRatio <= 1.0 && pageRatio > 0),
         assert(scaleRatio <= 1.0 && scaleRatio > 0),
@@ -42,6 +43,8 @@ class ScalePageView extends StatefulWidget {
   final double scaleRatio;
   final double opacityRatio;
   final double paddingTB;
+  final ScrollPhysics physics;
+
 
   @override
   State<StatefulWidget> createState() {
@@ -54,6 +57,13 @@ class ScalePageViewState extends State<ScalePageView> {
   ValueNotifier<double> selectedIndex = new ValueNotifier(0.0);
   List<Widget> backgrounds;
 
+  /// 在使用Key调用该处的时候,切记不能在最外层使用StatelessWidget
+  /// 这样可能会时Key发生变化,然后造成State重建
+  /// The purpose of a stateful widget is that they don't recreate their state object every time
+  /// they are built - otherwise they would be the same as a regular Widget.
+  /// A state object is recreated if the widget's runtimeType or key is different.
+  /// Otherwise the lifecycle didUpdateWidget is used to notify the State object that the configuration changed.
+  @Deprecated('This will lead to bugs.')
   void jumpToWithoutSettling(double value) {
     if (_pageController.position.pixels != value) {
       _pageController.position.jumpToWithoutSettling(value);
@@ -124,6 +134,7 @@ class ScalePageViewState extends State<ScalePageView> {
           controller: _pageController,
           children: _allItems(widget.children),
           onPageChanged: widget.onPageChange,
+          physics: widget.physics,
         ));
   }
 }