+++ CHANGELOG.md
@@ -6,3 +6,11 @@
 ## [1.0.0] - Official release.
 
 No Change
+
+## [1.0.1] - (Minor edits) Add Property
+
+add ScrollPhysics property
+
+## [1.0.2] - (Minor edits) Fix bug
+
+fix example recreate bug